+++ packages/source-preview-plugin/eslint.config.mjs
@@ -1,3 +1,8 @@
 import config from '@mdxeditor/tooling/eslint.config.mjs';
 
-export default config;
+export default [
+  ...config,
+  {
+    ignores: ['.ladle/**'],
+  },
+];