+++ packages/floating-selection-ui-plugin/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @mdxeditor/floating-selection-ui-plugin
 
+## 0.1.3
+
+### Patch Changes
+
+- 023f72f: remove console.log
+
+## 0.1.2
+
+### Patch Changes
+
+- fdc5984: Support read-only mode
+
+## 0.1.1
+
+### Patch Changes
+
+- aa2355e: Upgrade to latest mdxeditor dep
+
 ## 0.1.0
 
 ### Minor Changes