+++ packages/source-preview-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mdxeditor/source-preview-plugin
 
+## 0.2.0
+
+### Minor Changes
+
+- 821b460: Diff editor support
+
 ## 0.1.4
 
 ### Patch Changes