+++ packages/typeahead-plugin/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @mdxeditor/typeahead-plugin
 
+## 0.4.0
+
+### Minor Changes
+
+- 4375fec: Support completely custom editor for typeahead nodes.
+- e390a04: - Auto-support for dark theming (inherited from the MDXEditor CSS vars)
+  - Support for loading state when fetching typeahead options asynchronously
+  - Support for no results state when no options are available
+  - Show error state when fetching options fails
+
 ## 0.3.0
 
 ### Minor Changes